refactor(plot): rename message_flow.py into message_flow_old.py
